Move UsersTable
