+++ site/src/pages/UsersPage/UsersTable/UsersTableBody.tsx
@@ -6,15 +6,15 @@ import { ChooseOne, Cond } from "components/Conditionals/ChooseOne"
 import { Pill } from "components/Pill/Pill"
 import { FC } from "react"
 import { useTranslation } from "react-i18next"
-import * as TypesGen from "../../api/typesGenerated"
-import { combineClasses } from "../../utils/combineClasses"
-import { AvatarData } from "../AvatarData/AvatarData"
-import { EmptyState } from "../EmptyState/EmptyState"
+import * as TypesGen from "../../../api/typesGenerated"
+import { combineClasses } from "../../../utils/combineClasses"
+import { AvatarData } from "../../../components/AvatarData/AvatarData"
+import { EmptyState } from "../../../components/EmptyState/EmptyState"
 import {
   TableLoaderSkeleton,
   TableRowSkeleton,
-} from "../TableLoader/TableLoader"
-import { TableRowMenu } from "../TableRowMenu/TableRowMenu"
+} from "../../../components/TableLoader/TableLoader"
+import { TableRowMenu } from "../../../components/TableRowMenu/TableRowMenu"
 import { EditRolesButton } from "./EditRolesButton"
 import { Stack } from "components/Stack/Stack"
 import { EnterpriseBadge } from "components/DeploySettingsLayout/Badges"
